✨ Separate Logic, Protect User Endpoints, Add Me Endpoint
