+++ middlewares/jwt_validation.js
@@ -3,8 +3,9 @@ const jwt = require('jsonwebtoken')
 function validateToken (secret) {
   return (req, res, next) => {
     const { authorization } = req.headers
+    console.log(authorization)
     if (!authorization || !authorization.startsWith('Bearer ')) {
-      res.sendStatus(401)
+      res.sendStatus(400)
       return
     }
     const token = authorization.slice(7)
